+++ src/DataFixtures/UserFixtures.php
@@ -7,6 +7,8 @@ namespace App\DataFixtures;
 use App\Entity\User;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 /**
@@ -16,10 +18,22 @@ final class UserFixtures extends Fixture
 {
     public function __construct(
         private readonly UserPasswordHasherInterface $passwordHasher,
-    ) {}
+        #[Autowire('%kernel.environment%')]
+        private readonly string $environment,
+        private readonly LoggerInterface $logger,
+    ) {
+    }
 
     public function load(ObjectManager $manager): void
     {
+        if ('test' !== $this->environment) {
+            $this->logger->warning('Les fixtures ne doivent être chargées qu\'en environnement de test. Environnement actuel: {env}', [
+                'env' => $this->environment,
+            ]);
+
+            return;
+        }
+
         $user = new User();
         $user->setEmail('test@example.com');
         $user->setPassword($this->passwordHasher->hashPassword($user, 'password'));